Update project dependencies, enable ES modules and add mock data generator
In this commit, several changes were made to shift the project towards using JavaScript's ECMAScript modules and TypeScript. Added a "type" field in package.json with a value of "module", indicating that .js files should be treated as ES modules. Also changed the "module" field in tsconfig.json from "NodeNext" to "ESNext".

Added new dependencies, "@types/uuid" and "@faker-js/faker", to support the creation of a new file, "_fakerFabric.ts