Refactor code and update ECMAScript target version
The '_fakerFabric.ts' was removed and its contents were moved to 'unread.ts' for better organization and readability. Additionally, an HTTP request file (hands.http) and an environment file (http-client.env.json) were added for easier testing and configuration, respectively. The rewrites rule in 'vercel.json' was updated to make the path dynamic. The ECMAScript target version in 'tsconfig.json' was also updated from ESNext to ES2022 for compatibility with updated JS features.